feat: id의 int값을 long으로 변경

+++ android/Signal/app/src/main/java/com/ongo/signal/data/model/main/BoardRequestDTO.kt
@@ -1,10 +1,10 @@
 package com.ongo.signal.data.model.main
 
 data class BoardRequestDTO(
-    val userId: Int,
+    val userId: Long,
     val writer: String,
     val title: String,
     val content: String,
-    val type: Int,
+    val type: Long,
     val tags: List<TagDTO>
 )
