+++ android/Signal/app/src/main/java/com/ongo/signal/data/model/main/CommentRequestDTO.kt
@@ -1,7 +1,7 @@
 package com.ongo.signal.data.model.main
 
 data class CommentRequestDTO(
-    val boardId: Int,
-    val userId: Int,
+    val boardId: Long,
+    val userId: Long,
     val content: String
 )
